feat(GenericModel): Add trust_remote_code support
This commit introduces the `trust_remote_code` parameter to the `GenericModel`
and `GenericEngine` classes. This allows users to specify if they trust the
remote code when initializing a `GenericModel`.

The `trust_remote_code` parameter is passed through to the `CausalEngine` class,
which uses it when loading the model. This enhancement increases the flexibility
of the `GenericModel` class. It now can be used with models that require the
execution of remote code.

BREAKING CHANGE: This change should be used with caution. Trusting remote code
can have security implications. Always ensure that the source of the model is
trusted and the code is understood before enabling this option.